UPDATE: Forgot to push previously, so  actual progress is below and previous comment is below that.
This is the first "working" version of the app. A few more hours of work
and I will be ready for simple evaluation. I still need to give the user
the ability to choose photos from their sd card. I also need to fix the
bitmap out of memory error.

This version is the beginning of the camera module.
Everything is finished with it, except allowing the user to input
their own custom name for the photo.
I also found a bitmap big that I will need to work out.